+++ pipeline-maven/src/test/java/org/jenkinsci/plugins/pipeline/maven/docker/JavaGitContainerTest.java
@@ -28,15 +28,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.jenkinsci.plugins.pipeline.maven.AbstractIntegrationTest;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.EnabledIf;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-@EnabledIf(
-        value = "org.jenkinsci.plugins.pipeline.maven.util.Conditions#isLinuxAndDockerSocketExists",
-        disabledReason = "Needs Docker and Docker does not work on Windows 2019 servers CI agents")
-@Testcontainers
+@Testcontainers(disabledWithoutDocker = true)
 public class JavaGitContainerTest extends AbstractIntegrationTest {
 
     @Container
